refactor: move register to its own module
