+++ cmd/backend/internal/router/router.go
@@ -44,7 +44,7 @@ func (r *Router) RegisterRoutes(app *backend.App) {
 	r.GET("/users", app.UserHandler.GetAllUsers)
 	r.DELETE("/users/:id", app.UserHandler.DeleteUserById)
 
-	r.POST("/register", app.UserHandler.Register)
+	r.POST("/register", app.AuthHandler.Register)
 
 	r.GET("/usergroups", app.UserGroupHandler.GetAllUserGroups)
 	r.POST("/usergroups", app.UserGroupHandler.CreateUserGroup)
