+++ internal/auth/dto.go
@@ -1,0 +1,1 @@
+package auth
