+++ internal/user/service.go
@@ -6,11 +6,8 @@ import (
 
 	"github.com/kappusuton-yon-tebaru/backend/internal/logger"
 	"github.com/kappusuton-yon-tebaru/backend/internal/models"
-	"github.com/kappusuton-yon-tebaru/backend/internal/utils"
 	"github.com/kappusuton-yon-tebaru/backend/internal/werror"
 	"go.mongodb.org/mongo-driver/v2/bson"
-	"go.mongodb.org/mongo-driver/v2/mongo"
-	"go.uber.org/zap"
 )
 
 type Service struct {
@@ -23,26 +20,6 @@ func NewService(repo *Repository, logger *logger.Logger) *Service {
 		repo,
 		logger,
 	}
-}
-
-func (s *Service) Register(ctx context.Context, dto UserCredentialDTO) *werror.WError {
-	hashedPassword, err := utils.HashPassword(dto.Password)
-	if err != nil {
-		s.logger.Error("error occured while hashing password", zap.Error(err))
-		return werror.NewFromError(err).SetMessage("error occured while registering")
-	}
-
-	dto.Password = hashedPassword
-
-	_, err = s.repo.CreateUser(ctx, dto)
-	if err != nil && mongo.IsDuplicateKeyError(err) {
-		return werror.NewFromError(err).SetMessage("user already exists").SetCode(400)
-	} else if err != nil {
-		s.logger.Error("error occured while creating user", zap.Error(err))
-		return werror.NewFromError(err).SetMessage("error occured while registering")
-	}
-
-	return nil
 }
 
 func (s *Service) GetAllUsers(ctx context.Context) ([]models.User, error) {
